flask-url-shortner/requirements.txt
python-dotenv==0.10.3
flask-sqlalchemy
